Use the "Internet Code Page" property
This tells us the codepage which a text or HTML body is in, so we try to
map to a suitable charset parameter.  If not specified we now use
CP1252, which appears to be the correct default.  Previously we always
used charset="UTF-8" for text/plain and didn't specify a charset for
text/html.