Ensure body strings are always strings and not octet streams
When setting the plain text body from the corresponding property, a
plain text body encoded using UTF-16LE will be a real string consisting
of characters, while a plain text body encoded using the 8BIT encoding
is a sequence of bytes.

This means that for the later case the body must be encoded to a string
using the given code page.

A more durable solution would improve the property decoding method to
use the correct code page, but this is a bigger change.